fix: relocate injection tokens to config-provider module and update related imports
