+++ packages/frameworks/vue/src/config-provider/index.ts
@@ -1,3 +1,3 @@
 export { default as GenuiConfigProvider } from './ConfigProvider.vue';
 export * from './use-media-theme';
-export * from '../shared/injection-tokens';
+export * from './injection-tokens.js';
